knotx-template-engine module rename.
