+++ documentation/src/main/wiki/GettingStarted.md
@@ -86,6 +86,11 @@ Here's how JSON configuration files look:
 ```json
 {
   "verticles" : {  
+      "com.cognifide.knotx.server.KnotxServerVerticle": {
+         "config" : {
+            ...
+         }
+      },
       "com.cognifide.knotx.repository.HttpRepositoryVerticle": {
         "config" : {
             ...
@@ -96,22 +101,27 @@ Here's how JSON configuration files look:
             ...
         }
       },      
-      "com.cognifide.knotx.viewengine.ViewEngineVerticle": {
+      "com.cognifide.knotx.engine.view.ViewEngineVerticle": {
         "config" : {
             ...
         }
       },
-      "com.cognifide.knotx.server.KnotxServerVerticle": {
+      "com.cognifide.knotx.core.serviceadapter.http.HttpServiceAdapterVerticle" :{
         "config" : {
             ...
         }
-      },  
+      },
       "com.cognifide.knotx.mocks.MockRemoteRepositoryVerticle": {
         "config" : {
             ...
         }
       },
       "com.cognifide.knotx.mocks.MockServiceVerticle": {
+        "config" : {
+            ...
+        }
+      },
+      "com.cognifide.knotx.mocks.MockServiceAdapterVerticle": {
         "config" : {
             ...
         }
@@ -169,7 +179,7 @@ The config node consists of:
 #### 1.3. Engine section
 ```json
   ...
-  "com.cognifide.knotx.viewengine.ViewEngineVerticle": {
+  "com.cognifide.knotx.engine.view.ViewEngineVerticle": {
     "config": {
       "address": "knotx.core.viewengine",
       "template.debug": true,
@@ -249,7 +259,7 @@ This section configures the Knot.x HTTP server. The config node consists of:
 - **http.port** - an HTTP port on which the server listens for requests,
 - **allowed.response.headers** - list of the headers that should be passed back to the client
 - **repositories** - configuration of repositories. It's a array of mappings what paths are supported by what repository verticles (by specifing its event bus addresses). The order of mappings is important as they are evaluated from top to down on each request. The first one matched will handle the request or, if no repository is matched, **Knot.x** will return a `404 Not found` response for the given request.
-- **engine** - configuration about templating engine dependency. You can configure event bus **address** of engine verticle here.
+- **engine** - configuration about view engine dependency. You can configure event bus **address** of engine verticle here.
 
 #### 1.5 MockRepo section
 ```json
