+++ documentation/src/main/wiki/WhatsNew.md
@@ -1,9 +1,3 @@
-# What's new in 0.4.0
+# What's new in 0.6.0
 
-* Comprehensive verticles deployment options
-* Placeholders in service url support
-* Repositories request / response HTTP headers rules
-* Services request / response HTTP headers rules
-* Caching service calls within one request
-* Custom handlebars helpers registration support
-* Specialized repositories verticle implementation
+*
