+++ knotx-core/knotx-view-engine/src/main/java/com/cognifide/knotx/engine/view/ViewEngineVerticle.java
@@ -15,11 +15,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.cognifide.knotx.viewengine;
+package com.cognifide.knotx.engine.view;
 
 import com.cognifide.knotx.dataobjects.RenderRequest;
 import com.cognifide.knotx.dataobjects.RenderResponse;
-import com.cognifide.knotx.viewengine.impl.TemplateEngine;
+import com.cognifide.knotx.engine.view.impl.TemplateEngine;
 
 import io.vertx.core.Context;
 import io.vertx.core.Vertx;
