+++ knotx-core/knotx-view-engine/src/main/java/com/cognifide/knotx/engine/view/impl/DefaultHandlebarsHelpers.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.cognifide.knotx.viewengine.impl;
+package com.cognifide.knotx.engine.view.impl;
 
 import com.github.jknack.handlebars.Handlebars;
 import com.github.jknack.handlebars.Helper;
