+++ knotx-core/knotx-view-engine/src/main/java/com/cognifide/knotx/engine/view/parser/HtmlParser.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.cognifide.knotx.viewengine.parser;
+package com.cognifide.knotx.engine.view.parser;
 
 import com.google.common.collect.Lists;
