+++ knotx-example/knotx-example-monolith/src/main/resources/knotx-example-monolith.json
@@ -78,7 +78,7 @@
         "catalogue": ""
       }
     },
-    "com.cognifide.knotx.viewengine.ViewEngineVerticle": {
+    "com.cognifide.knotx.engine.view.ViewEngineVerticle": {
       "config": {
         "address": "knotx.core.viewengine",
         "template.debug": true,
