+++ knotx-example/knotx-example-monolith/src/main/resources/logback.xml
@@ -33,7 +33,7 @@
     <appender-ref ref="STDOUT"/>
   </logger>
 
-  <logger name="com.cognifide.knotx.viewengine" level="TRACE"
+  <logger name="com.cognifide.knotx.engine.view" level="TRACE"
           additivity="false">
     <appender-ref ref="STDOUT"/>
   </logger>
